requirements.txt
google-generativeai
pystray
pynput
pyperclip
plyer
Pillow
